pyproject.toml
[tool.poetry]
name = "duplicate-stickers-remover-bot"
version = "0.1.0"
description = ""
authors = ["isayakhov <isayakhov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.4"
python-telegram-bot = "^13.10"
"backports.zoneinfo" = "^0.2.1"
ImageHash = "^4.2.1"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pre-commit = "^2.17.0"
mypy = "^0.931"
pylint = "^2.12.2"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-faker = "^2.0.0"
pytest-mock = "^3.7.0"
pytest-factoryboy = "^2.1.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119
multi_line_output = 3
default_section = "THIRDPARTY"
known_first_party = "app"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
forced_separate = "tests"
skip_glob = "**/migrations/*.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
